+++ source/global/PointsForPlacesPreset.h
@@ -11,13 +11,13 @@ public:
     PointsForPlacesPreset(QString name = "");
 private:
     QString name;
-    QMap<int, int> pointsForPlaces; //1. Miejsce, 2. Punkty
+    QMap<int, double> pointsForPlaces; //1. Miejsce, 2. Punkty
 public:
     QString getName() const;
     void setName(const QString &newName);
-    QMap<int, int> getPointsForPlaces() const;
-    QMap<int, int> & getPointsForPlacesReference();
-    void setPointsForPlaces(const QMap<int, int> &newPointsForPlaces);
+    QMap<int, double> getPointsForPlaces() const;
+    QMap<int, double> & getPointsForPlacesReference();
+    void setPointsForPlaces(const QMap<int, double> &newPointsForPlaces);
 public:
     static QJsonObject getPointsForPlacesPresetJsonObject(PointsForPlacesPreset & preset);
     static QVector<PointsForPlacesPreset> getPointsForPlacesPresetsVectorFromJson(const QByteArray & bytes);
